account Manager
Load Accounts into an Account Manager instead of having the functionality in the gui

Changed up the Account name variables to remove account (it's in the Account Class ya dummy)
